# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "C08 Document Template API",
        "title": "Document Template API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "nguyenhai.hoang1@etc.vn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/apikey": {
            "post": {
                "description": "API generate apikey",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "API generate apikey",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.GenerateApiKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.GenerateApiKeyResponse"
                        }
                    }
                }
            }
        },
        "/v1/template/docx/render": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "API generate docx template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Template"
                ],
                "summary": "API generate docx template",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.DocxRequestGenerate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "vo.DocxRequestGenerate": {
            "type": "object",
            "required": [
                "parameters",
                "template",
                "type"
            ],
            "properties": {
                "parameters": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "template": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "pdf",
                        "docx"
                    ]
                }
            }
        },
        "vo.GenerateApiKeyRequest": {
            "type": "object",
            "required": [
                "accessKeyID",
                "secretAccessKey"
            ],
            "properties": {
                "accessKeyID": {
                    "type": "string"
                },
                "secretAccessKey": {
                    "type": "string"
                }
            }
        },
        "vo.GenerateApiKeyResponse": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}